Optimized code.
